Improve the building of the docker images and document them
The images now listen on 8081 to not require root permissions.
Since AFAIK they are built for testing/demo they should not need root.

The generate scripts now use the built module and not the system installed one.

The generate scripts and the images now use a compatible configuration as the top-level example.