ci: Pin nethsm image to v2.2 (3e45f2f3)
The latest NetHSM test versions removed support for EC_P224 but it is
unclear if this will also apply for the next stable release.  Thus this
patch pins the nethsm image used in the CI to the latest stable release
v2.2 until it is decided if EC_P224 is removed for good.

See also:
https://github.com/Nitrokey/nethsm/releases/tag/v2.2
https://github.com/Nitrokey/nethsm/issues/15