ci: Test module execution on Windows and macOS
This patch adds simple smoke tests to make sure that the compiled
pkcs11 modules work with pkcs11-tool.  As we cannot use service
containers with windows and macos runners, we use the public demo
instance instead.  As we don’t execute any actions for the tests, this
should be fine.