api: Use wrappers for stateful API functions
The decrypt, encrypt and sign API functions maintain state that has to
be cleared after function calls depending on the result of the function
call. Currently, this is implemented manually on every possible return
path, but that is hard to maintain and check. This patch introduces
wrapper functions that clear the state depending on the result of the
actual implementation and according to the requirements of the
specification.