Reject invalid IDs when creating or changing objects
Previously, we replaced invalid IDs with their hex encoding internally.
This adds complexity while it is not clear that this feature is really
needed. With this patch, we just reject invalid IDs. This patch also
introduces an Id type to unify the validation logic in one place.

See also: https://github.com/Nitrokey/nethsm-pkcs11/issues/273